Refactored build tasks
