+++ package.json
@@ -59,6 +59,7 @@
     "handlebars-helpers": "~0.5.5",
     "jasmine-fixture": "~1.0.8",
     "jshint-stylish": "~0.1.3",
+    "load-grunt-config": "^0.16.0",
     "lodash": "^2.4.1",
     "seed-random": "~2.2.0",
     "time-grunt": "^1.0.0"
