+++ tasks/imagemin.js
@@ -1,0 +1,25 @@
+/**
+ * imagemin
+ * ========
+ *
+ * The following *-min tasks produce minified files in the dist folder.
+ *
+ * Link: https://github.com/gruntjs/grunt-contrib-imagemin
+ */
+
+'use strict';
+
+module.exports = function (grunt) {
+  return {
+    dist: {
+      files: [
+        {
+          expand: true,
+          cwd: '<%= pkg.config.source %>/images',
+          src: '{,*/}*.{png,jpg,jpeg,gif}',
+          dest: '<%= pkg.config.dist %>/images'
+        }
+      ]
+    }
+  }
+};
